chore: remove some useless types, refactor some codes into `MarkdownPreview.vue` component, update HoverPreview into Array
